zippy for debug
